requirements.txt
flask==3.0.3
twilio==9.0.5
langgraph==0.0.39
langchain==0.1.20
langchain-anthropic==0.1.13
anthropic>=0.86.0
playwright==1.44.0
sqlalchemy==2.0.30
python-dotenv==1.0.1
psycopg2-binary==2.9.9
fpdf2==2.7.9
tavily-python==0.3.3
msal==1.28.0
msgraph-sdk==1.1.0
requests==2.31.0
